requirements.txt
requests
python-dotenv
langchain-core
langchain-google-genai
langchain
todoist-api-python
pydantic
tzdata